+++ Carbon.Security/Functions/Grant-CPrivilege.ps1
@@ -3,15 +3,20 @@ function Grant-CPrivilege
 {
     <#
     .SYNOPSIS
-    Grants an identity priveleges to perform system operations.
+    Grants an account privileges to perform system operations.
 
     .DESCRIPTION
-    *Privilege names are **case-sensitive**.* Valid privileges are documented on Microsoft's website: [Privilege Constants](http://msdn.microsoft.com/en-us/library/windows/desktop/bb530716.aspx) and [Account Right Constants](http://msdn.microsoft.com/en-us/library/windows/desktop/bb545671.aspx). Here is the most current list, as of August 2014:
+    The `Grant-CPrivilege` function grants a user/group rights and privileges. Pass the name of the user/group to the
+    `Identity` parameter. Pass the list of account rights and/or privileges to grant to the `Privilege` parameter. The
+    account is granted any rights/privileges it doesn't currently have.
+
+    Rights and privilege names are documented on Microsoft's website, duplicated below. These lists may be out-of-date.
+
+    [Privilege Constants](https://learn.microsoft.com/en-us/windows/win32/secauthz/privilege-constants):
 
      * SeAssignPrimaryTokenPrivilege
      * SeAuditPrivilege
      * SeBackupPrivilege
-     * SeBatchLogonRight
      * SeChangeNotifyPrivilege
      * SeCreateGlobalPrivilege
      * SeCreatePagefilePrivilege
@@ -19,29 +24,22 @@ function Grant-CPrivilege
      * SeCreateSymbolicLinkPrivilege
      * SeCreateTokenPrivilege
      * SeDebugPrivilege
-     * SeDenyBatchLogonRight
-     * SeDenyInteractiveLogonRight
-     * SeDenyNetworkLogonRight
-     * SeDenyRemoteInteractiveLogonRight
-     * SeDenyServiceLogonRight
+     * SeDelegateSessionUserImpersonatePrivilege
      * SeEnableDelegationPrivilege
      * SeImpersonatePrivilege
      * SeIncreaseBasePriorityPrivilege
      * SeIncreaseQuotaPrivilege
      * SeIncreaseWorkingSetPrivilege
-     * SeInteractiveLogonRight
      * SeLoadDriverPrivilege
      * SeLockMemoryPrivilege
      * SeMachineAccountPrivilege
      * SeManageVolumePrivilege
-     * SeNetworkLogonRight
      * SeProfileSingleProcessPrivilege
      * SeRelabelPrivilege
      * SeRemoteInteractiveLogonRight
      * SeRemoteShutdownPrivilege
      * SeRestorePrivilege
      * SeSecurityPrivilege
-     * SeServiceLogonRight
      * SeShutdownPrivilege
      * SeSyncAgentPrivilege
      * SeSystemEnvironmentPrivilege
@@ -54,6 +52,18 @@ function Grant-CPrivilege
      * SeUndockPrivilege
      * SeUnsolicitedInputPrivilege
 
+    [Account Right Constants](https://learn.microsoft.com/en-us/windows/win32/secauthz/account-rights-constants):
+
+     * SeBatchLogonRight
+     * SeDenyBatchLogonRight
+     * SeDenyInteractiveLogonRight
+     * SeDenyNetworkLogonRight
+     * SeDenyRemoteInteractiveLogonRight
+     * SeDenyServiceLogonRight
+     * SeInteractiveLogonRight
+     * SeNetworkLogonRight
+     * SeServiceLogonRight
+
     .LINK
     Get-CPrivilege
 
@@ -64,70 +74,89 @@ function Grant-CPrivilege
     Test-CPrivilege
 
     .LINK
-    http://msdn.microsoft.com/en-us/library/windows/desktop/bb530716.aspx
+    Test-CPrivilegeName
 
     .LINK
-    http://msdn.microsoft.com/en-us/library/windows/desktop/bb545671.aspx
+    https://learn.microsoft.com/en-us/windows/win32/secauthz/privilege-constants
+
+    .LINK
+    https://learn.microsoft.com/en-us/windows/win32/secauthz/account-rights-constants
 
     .EXAMPLE
     Grant-CPrivilege -Identity Batcomputer -Privilege SeServiceLogonRight
 
-    Grants the Batcomputer account the ability to logon as a service. *Privilege names are **case-sensitive**.*
+    Grants the Batcomputer account the ability to logon as a service.
     #>
     [CmdletBinding()]
     param(
-        [Parameter(Mandatory=$true)]
-        [string]
-        # The identity to grant a privilege.
-        $Identity,
+        # The user/group name to grant rights/privileges.
+        [Parameter(Mandatory)]
+        [String] $Identity,
 
-        [Parameter(Mandatory=$true)]
-        [string[]]
-        # The privileges to grant. *Privilege names are **case-sensitive**.*
-        $Privilege
+        # The rights/privileges to grant.
+        #
+        # [Privilege names are documented on the "Privilege Constants"
+        # page.](https://learn.microsoft.com/en-us/windows/win32/secauthz/privilege-constants)
+        #
+        # [Rights names are documented on the "Account Rights Constants"
+        # page.](https://learn.microsoft.com/en-us/windows/win32/secauthz/account-rights-constants)
+        [Parameter(Mandatory)]
+        [String[]] $Privilege
     )
 
     Set-StrictMode -Version 'Latest'
-
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    $account = Resolve-CIdentity -Name $Identity -NoWarn
+    $account = Resolve-CIdentity -Name $Identity
     if( -not $account )
+    {
+        return
+    }
+
+    $privilegesToGrant = $Privilege | Where-Object { -not (Test-CPrivilege -Identity $account.FullName -Privilege $_) }
+    if (-not $privilegesToGrant)
+    {
+        return
+    }
+
+    $unknownPrivileges = $privilegesToGrant | Where-Object { -not (Test-CPrivilegeName -Name $_) }
+    if ($unknownPrivileges)
+    {
+        $privileges = 'privilege'
+        $thatThose = 'that'
+        $isAre = 'is'
+        if (($unknownPrivileges | Measure-Object).Count -gt 1)
+        {
+            $privileges = 'privileges'
+            $thatThose = 'those'
+            $isAre = 'are'
+        }
+        $msg = "Failed to grant the $($account.FullName) account $($unknownPrivileges -join ', ') ${privileges} " +
+               "because ${thatThose} ${privileges} ${isAre} unknown."
+        Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+    }
+
+    # Privilege names are case-sensitive when granting, so get the actual value of the privilege names.
+    $privilegesToGrant = $privilegesToGrant | Test-CPrivilegeName -PassThru | Where-Object { $_ }
+    if (-not $privilegesToGrant)
+    {
+        return
+    }
+
+    $pHandle = Invoke-AdvApiLsaOpenPolicy -DesiredAccess CreateAccount,LookupNames
+    if (-not $pHandle)
     {
         return
     }
 
     try
     {
-        [Carbon.Security.Privilege]::GrantPrivileges( $account.FullName, $Privilege )
+        Write-Information "$($account.FullName)  + $($privilegesToGrant -join ',')"
+        Invoke-AdvApiLsaAddAccountRights -PolicyHandle $pHandle -Sid $account.Sid -Privilege $privilegesToGrant |
+            Out-Null
     }
-    catch
+    finally
     {
-        $ex = $_.Exception
-        do
-        {
-            if( $ex -is [ComponentModel.Win32Exception] -and $ex.Message -eq 'No such privilege. Indicates a specified privilege does not exist.' )
-            {
-                $msg = 'Failed to grant {0} {1} privilege(s): {2}  *Privilege names are **case-sensitive**.*' -f `
-                        $account.FullName,($Privilege -join ','),$ex.Message
-                Write-Error -Message $msg
-                return
-            }
-            else
-            {
-                $ex = $ex.InnerException
-            }
-        }
-        while( $ex )
-
-        $ex = $_.Exception
-        Write-Error -Message ('Failed to grant {0} {1} privilege(s): {2}' -f $account.FullName,($Privilege -join ', '),$ex.Message)
-
-        while( $ex.InnerException )
-        {
-            $ex = $ex.InnerException
-            Write-Error -Exception $ex
-        }
+        Invoke-AdvApiLsaClose -PolicyHandle $pHandle | Out-Null
     }
 }
-