+++ Carbon.Security/Functions/Test-CPrivilege.ps1
@@ -4,7 +4,7 @@ function Test-CPrivilege
     <#
     .SYNOPSIS
     Tests if an identity has a given privilege.
-    
+
     .DESCRIPTION
     Returns `true` if an identity has a privilege.  `False` otherwise.
 
@@ -19,31 +19,30 @@ function Test-CPrivilege
 
     .LINK
     Revoke-CPrivilege
-    
+
+    .LINK
+    Test-CPrivilegeName
+
     .EXAMPLE
     Test-CPrivilege -Identity Forrester -Privilege SeServiceLogonRight
-    
+
     Tests if `Forrester` has the `SeServiceLogonRight` privilege.
     #>
     [CmdletBinding()]
     param(
-        [Parameter(Mandatory=$true)]
-        [string]
         # The identity whose privileges to check.
-        $Identity,
-        
-        [Parameter(Mandatory=$true)]
-        [string]
-        # The privilege to check.
-        $Privilege
-    )
-    
-    Set-StrictMode -Version 'Latest'
+        [Parameter(Mandatory)]
+        [String] $Identity,
 
+        # The privilege to check.
+        [Parameter(Mandatory)]
+        [String] $Privilege
+    )
+
+    Set-StrictMode -Version 'Latest'
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    $matchingPrivilege = Get-CPrivilege -Identity $Identity |
-                            Where-Object { $_ -eq $Privilege }
-    return ($matchingPrivilege -ne $null)
+    $matchingPrivilege = Get-CPrivilege -Identity $Identity | Where-Object { $_ -eq $Privilege }
+    return ($null -ne $matchingPrivilege)
 }